requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4      
async-timeout==3.0.1
attrs==21.2.0       
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
discord.py==1.7.2
distlib==0.3.2
filelock==3.0.12
idna==2.10
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
yarl==1.6.3